+++ src/presentation/web/app/api/applications/[id]/cloud-deploy/status/route.ts
@@ -8,7 +8,7 @@ import type { NextRequest } from 'next/server';
 import { NextResponse } from 'next/server';
 import { resolve } from '@/lib/server-container';
 import type { GetCloudDeploymentStatusUseCase } from '@shepai/core/application/use-cases/cloud-deploy/get-cloud-deployment-status.use-case';
-import { ApplicationNotFoundError } from '@shepai/core/application/use-cases/cloud-deploy/select-cloud-provider.use-case';
+import { ApplicationNotFoundError } from '@shepai/core/domain/errors/application-not-found.error';
 
 export const dynamic = 'force-dynamic';
